requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.29
websockets>=12.0
asterix-decoder>=0.7.11
